requirements.txt
Flask
torchvision
torch
pillow